Build Setting Configurator code
Apparently Unity saves its build settings via the Library, too bad I need to delete the Library to submit the zip file. This code is made primarily with the help of AI, thank god on my end. Anyways this code just sets up the build automatically so you don't have to manually do it when you start the project.